requirements.txt
numpy==1.23.2
pygame==2.1.2
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
